Added goal of growing volunteer base
I believe the purpose of BOINC is to perform grid computing. As stated today, the purpose is just for the software to function well - reliably do the computing over time. There is no goal of increasing the size of the grid. Everything that BOINC is and achieves is multiplied by the number of volunteers. With twice the volunteers every single other action by every person involved in BOINC in any way has twice the impact on the amount of grid computing performed, twice the impact on the world.

I don't know how much growth is possible or exactly how to achieve every part of growing the volunteer base, but it does seem necessary to at least state that growing the number of volunteers is a goal, if the overall goal is to perform the maximum amount of grid computing.

Thanks very much for letting me comment.
Quentin

(I am aware of Grid Republic, Charity Engine etc. I'm just talking about stating the goal of growth)